Revert "Feat: CalendarView TCA화 및 배경과 화면 제목 추가"
